mage-zoomcamp/requirements.txt
requests
python-dotenv
matplotlib
plotly
nasdaq-data-link